+++ SignalServiceKit/src/Messages/TSGroupModel.m
@@ -8,7 +8,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-const NSUInteger kGroupIdLength = 16;
+const int32_t kGroupIdLength = 16;
 
 @interface TSGroupModel ()
